requirements.txt
# Core
datasets>=2.18.0
pandas>=2.2.0
pyarrow>=15.0.0
pydantic>=2.0.0

# API (Phase 4)
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
python-dotenv>=1.0.0

# LLM (Phase 3)
groq>=0.9.0            # Groq inference API (primary)
openai>=1.30.0
google-generativeai>=0.5.0
httpx>=0.27.0          # Groq + Ollama adapters
json-repair>=0.28.0    # malformed JSON fallback parser

# Fuzzy matching (Phase 2)
rapidfuzz>=3.9.0

# Testing (Phase 6)
pytest>=8.0.0
pytest-cov>=5.0.0
pytest-asyncio>=0.23.0

# Frontend — Streamlit MVP (Phase 5)
streamlit>=1.35.0
requests>=2.31.0        # used by streamlit_app.py to call the FastAPI backend
